Move viewer store and effects to viewer/store folder
