+++ libs/search-widget/src/old-components/viewer/store/viewer-effects.ts
@@ -1,15 +1,15 @@
 import type { Subscription } from 'rxjs';
 import { concatMap, filter, switchMap, tap } from 'rxjs/operators';
-import { getResource, loadEntities } from '../api';
+import { getResource, loadEntities } from '../../../core/api';
 import type { Resource } from '@nuclia/core';
-import { resource } from './resource.store';
-import { isViewerOpen } from './modal.store';
-import { formatQueryKey, updateQueryParams } from '../utils';
-import { canEditLabels } from './widget.store';
-import { initLabelStore } from './effects';
-import { entityGroups } from './entities.store';
-import { displayedResource } from './search.store';
-import type { DisplayedResource } from '../models';
+import { resource } from '../../../core/stores/resource.store';
+import { isViewerOpen } from '../../../core/stores/modal.store';
+import { formatQueryKey, updateQueryParams } from '../../../core/utils';
+import { canEditLabels } from '../../../core/stores/widget.store';
+import { initLabelStore } from '../../../core/stores/effects';
+import { entityGroups } from '../../../core/stores/entities.store';
+import { displayedResource } from '../../../core/stores/search.store';
+import type { DisplayedResource } from '../../../core/models';
 
 const subscriptions: Subscription[] = [];
 const previewQueryKey = formatQueryKey('preview');
